Remove listener channel
