+++ lib/channels/route_channel.ex
@@ -43,6 +43,35 @@ defmodule ApplicationRunner.RouteChannel do
         {:error, ErrorHelpers.translate_error(BusinessError.invalid_channel_name())}
       end
 
+      ######
+      # IN #
+      ######
+
+      def handle_in("run", %{"code" => code, "event" => event}, socket) do
+        handle_run(socket, code, event)
+      end
+
+      def handle_in("run", %{"code" => code}, socket) do
+        handle_run(socket, code)
+      end
+
+      def handle_run(socket, code, event \\ %{}) do
+        session_id = Map.fetch!(socket.assigns, :session_id)
+
+        Logger.debug("Handle run #{code}")
+
+        case Session.send_client_event(session_id, code, event) do
+          {:error, err} ->
+            Phoenix.Channel.push(socket, "error", ErrorHelpers.translate_error(err))
+            {:reply, {:error, %{"error" => err}}, socket}
+
+          _ ->
+            {:reply, {:ok, %{}}, socket}
+        end
+
+        {:noreply, socket}
+      end
+
       ########
       # INFO #
       ########
